refactor(*): add structures and refactor
