+++ include/minishell.h
@@ -6,7 +6,7 @@
 /*   By: mgerbaud <mgerbaud@student.42nice.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/21 16:36:19 by mgerbaud          #+#    #+#             */
-/*   Updated: 2022/11/30 10:46:07 by aguiri           ###   ########.fr       */
+/*   Updated: 2022/12/01 18:40:01 by aguiri           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,25 +26,9 @@
 // Personal library
 
 # include "libft.h"
+
+# include "structures.h"
 # include "builtins.h"
-
-// ****************************************************************************
-// Structures
-
-typedef struct s_paths
-{
-	char			*str;
-	struct s_paths	*next;
-}					t_paths;
-
-typedef struct s_cmd
-{
-	char	*cmdpath;
-	char	*cmd;
-	char	*flags;
-	char	*arg;
-}			t_cmd;
-
 
 // ****************************************************************************
 // Functions
@@ -54,4 +38,4 @@ int		search_path(t_paths *hpaths, t_cmd *cmd, char *str);
 int		exec_cmd(t_cmd *cmd, char *av[], char *env[]);
 char	**ft_split(const char *str, char sep);
 
-#endif
+#endif // MINISHELL_H
